# Faster path for `onehotbatch(::CUArray{Int}, ::UnitRange)`