--- a/test/gpu.jl
+++ b/test/gpu.jl
@@ -32,8 +32,12 @@ end
   @test y1.indices == y2.indices
   @test_broken y1 == y2  # issue 28
 
-  @test_throws Exception onehotbatch([1, 3, 0, 2] |> cu, 1:10)
-  @test_throws Exception onehotbatch([1, 3, 0, 2] |> cu, -2:2)
+  if !CUDA.functional()
+    # Here CUDA gives an error which @test_throws does not notice,
+    # although with JLArrays @test_throws it's fine.
+    @test_throws Exception onehotbatch([1, 3, 0, 2] |> cu, 1:10)
+    @test_throws Exception onehotbatch([1, 3, 0, 2] |> cu, -2:2)
+  end
 end
 
 @testset "onecold gpu" begin